third_party/blink/web_tests/fast/dom/shadow/remove-and-insert-style-expected.txt

CONSOLE WARNING: line 48: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Test for Issue: 246300: Styles in nested shadows are not recalculated correctly on insertion.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
